python-backend/requirements.txt
# Core web framework
flask==2.3.3
flask-cors==4.0.0

# Image processing and computer vision
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scipy==1.11.3

# OCR and text processing
pytesseract==0.3.10
easyocr==1.7.0

# Machine learning and data analysis
pandas==2.0.3
matplotlib==3.7.2
scikit-image==0.21.0
scikit-learn==1.3.0

# Technical analysis
talib==0.4.28
mplfinance==0.12.10a0

# Additional utilities
requests==2.31.0
python-dotenv==1.0.0
base64==2.2.2